+++ flutter_local_notifications/lib/src/platform_specifics/android/notification_channel.dart
@@ -9,7 +9,7 @@ class AndroidNotificationChannel {
     this.id,
     this.name,
     this.description, {
-    this.importance = Importance.Default,
+    this.importance = Importance.defaultImportance,
     this.playSound = true,
     this.sound,
     this.enableVibration = true,